[VW-0] Remove advisories dead code
