--- src/features/advisories/hooks/use-advisories-params.ts
@@ -1,8 +1,0 @@
-"use client";
-
-import { useQueryStates } from "nuqs";
-import { advisoriesParams } from "../params";
-
-export const useAdvisoriesParams = () => {
-  return useQueryStates(advisoriesParams);
-};
