--- src/features/advisories/params.ts
@@ -1,3 +1,0 @@
-import { createPaginationParams } from "@/lib/url-state";
-
-export const advisoriesParams = createPaginationParams();
