--- src/features/advisories/server/params-loader.ts
@@ -1,4 +1,0 @@
-import { createLoader } from "nuqs/server";
-import { advisoriesParams } from "../params";
-
-export const advisoriesParamsLoader = createLoader(advisoriesParams);
