+++ src/features/tracking/components/tracking.test.tsx
@@ -269,7 +269,6 @@ const sampleTicket = (overrides: Record<string, unknown> = {}) => ({
     issues: 0,
     vulnerabilities: 0,
     remediations: 0,
-    advisories: 0,
     assets: 0,
   },
   isWatching: false,
@@ -560,7 +559,6 @@ describe("TrackingList", () => {
               issues: 0,
               vulnerabilities: 0,
               remediations: 0,
-              advisories: 0,
               assets: 0,
             },
           }),
@@ -589,7 +587,6 @@ describe("TrackingList", () => {
               issues: 0,
               vulnerabilities: 0,
               remediations: 0,
-              advisories: 0,
               assets: 0,
             },
           }),
